Remove NVK quote on the home page
I don't think it is a good idea to include a quote about the thread of government take-downs.

It's a sensitive communication issue, and I think it would be helpful to have a clear stance across the site and product. We can learn from the Tor project. For example, they have an [abuse FAQ](https://support.torproject.org/abuse/), which includes this line:

> We hate that there are some people who use Tor for nefarious purposes, and we condemn the misuse and exploitation of our technology for criminal activity. It's essential to understand that criminal intent lies with the individuals and not the tools they use.

The Tor [history/about page](https://www.torproject.org/about/history/) also does a good job at outlining the mission.

Probably a bigger discussion to be had, but for starters I think it might be a good idea to just remove that quote.